requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.4.4
